norm in run_commands
