+++ Makefile
@@ -13,7 +13,7 @@ SRCS	=	parser/main.c parser/preparser.c parser/preparser_utils.c parser/parser.c
 			parser/new_env_list.c \
 			parser/make_envp.c parser/parser_redirect.c parser/parser_fds.c parser/ft_parser_utils.c\
  			run_commands/handler_pipe.c build_in/build_in.c build_in/build_in2.c build_in/build_in3.c \
- 			run_commands/execute.c run_commands/handler_cmd.c run_commands/handler_signal.c
+ 			run_commands/execute.c run_commands/handler_cmd.c run_commands/cmd_check.c run_commands/handler_signal.c run_commands/change_fds.c
 
 OBJS	=	${SRCS:.c=.o}
