+++ minishell.h
@@ -116,4 +116,8 @@ void	ft_handler_child(int sig);
 void	ft_handler_main(int sig);
 void	ft_handler_ctrl_d(void);
 void	ft_handler_heredoc(int sig);
+
+//change fds
+void	ft_dup_fd_in(int reserved_stdin, t_fds *fds);
+void	ft_dup_fd_out(int reserved_stdout, t_fds *fds);
 #endif
