+++ run_commands/handler_pipe.c
@@ -1,51 +1,16 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   handler_pipe.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jcollin <jcollin@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/25 16:19:15 by jcollin           #+#    #+#             */
+/*   Updated: 2022/03/25 16:19:17 by jcollin          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../minishell.h"
-
-void	ft_free(char **dst)
-{
-	size_t	i;
-
-	i = 0;
-	while (dst[i])
-	{
-		free(dst[i]);
-		i++;
-	}
-	free(dst);
-}
-
-char	*ft_parsing_path(char *cmd, char **envp)
-{
-	char	**paths;
-	char	*cmdpath;
-	char	*path;
-	int		i;
-
-	i = 0;
-	if (cmd == NULL || ft_strchr(cmd, '/') != 0)
-		return (cmd);
-	while (envp[i] && envp[i] != ft_strnstr(envp[i], "PATH=", 5))
-		i++;
-	if ((ft_strnstr(envp[i], "PATH=", 5)) == NULL)
-		return ("command not found");
-	path = envp[i];
-	i = 0;
-	paths = ft_split(path + 5, ':');
-	while (paths[i])
-	{
-		path = ft_strjoin(paths[i], "/");
-		cmdpath = ft_strjoin(path, cmd);
-		free(path);
-		if (access(cmdpath, R_OK) == 0)
-		{
-			ft_free(paths);
-			return (cmdpath);
-		}
-		free(cmdpath);
-		i++;
-	}
-	ft_free(paths);
-	return ("command not found");
-}
 
 int	execve_for_pipe(t_list *elem)
 {
@@ -68,6 +33,19 @@ int	execve_for_pipe(t_list *elem)
 	}
 }
 
+static void	ft_child_proc(int *fd, t_fds *fds)
+{
+	usleep(100);
+	close(fd[1]);
+	dup2(fd[0], STDIN_FILENO);
+	if (fds->next->fd_in != 0)
+	{
+		dup2(fds->next->fd_in, STDIN_FILENO);
+		close(fds->next->fd_in);
+	}
+	close(fd[0]);
+}
+
 void	ft_launch_proc(t_list *elem, t_fds *fds)
 {
 	int		fd[2];
@@ -77,7 +55,6 @@ void	ft_launch_proc(t_list *elem, t_fds *fds)
 	{
 		perror("minishelchik :");
 		exit(1);
-		///free all
 	}
 	pid = fork();
 	if (pid == 0)
@@ -93,20 +70,10 @@ void	ft_launch_proc(t_list *elem, t_fds *fds)
 		execve_for_pipe(elem);
 	}
 	else
-	{
-		usleep(100);
-		close(fd[1]);
-		dup2(fd[0], STDIN_FILENO);
-		if (fds->next->fd_in != 0)
-		{
-			dup2(fds->next->fd_in, STDIN_FILENO);
-			close(fds->next->fd_in);
-		}
-		close(fd[0]);
-	}
+		ft_child_proc(fd, fds);
 }
 
-void ft_wait(pid_t l_pid, int reserved_stdin, int reserved_stdout)
+void	ft_wait(pid_t l_pid, int reserved_stdin, int reserved_stdout)
 {
 	pid_t	temp_pid;
 	int		status;
@@ -121,27 +88,6 @@ void ft_wait(pid_t l_pid, int reserved_stdin, int reserved_stdout)
 		if (temp_pid == l_pid)
 			g_var->state = WEXITSTATUS(status);
 		temp_pid = waitpid(-1, &status, 0);
-	}
-}
-
-static void	ft_dup_fd_in(int reserved_stdin, t_fds *fds)
-{
-	if (fds->fd_in != 0)
-	{
-		dup2(fds->fd_in, STDIN_FILENO);
-		close(fds->fd_in);
-	}
-	else
-		dup2(reserved_stdin, STDIN_FILENO);
-}
-
-static void	ft_dup_fd_out(int reserved_stdout, t_fds *fds)
-{
-	dup2(reserved_stdout, STDOUT_FILENO);
-	if (fds->fd_out != 0)
-	{
-		dup2(fds->fd_out, STDOUT_FILENO);
-		close(fds->fd_out);
 	}
 }
 
@@ -160,7 +106,6 @@ int	ft_exec_pipes(t_list *elem, t_fds *fds)
 	ft_dup_fd_in(reserved_stdin, tmp_fds);
 	while (tmp_fds->next != NULL)
 	{
-		// ft_init_signal_handler(ft_handler_child);
 		ft_launch_proc(tmp, tmp_fds);
 		tmp = tmp->next;
 		tmp_fds = tmp_fds->next;
